# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for learning English via dictation (video-based).",
        "title": "Dictation Learning System API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8002",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/login": {
            "post": {
                "description": "Xac thuc thong tin dang nhap va tra ve access token + refresh token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Dang nhap",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-familytree_internal_service.LoginInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dang nhap thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_service.TokenPair"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Vo hieu hoa access token hien tai (dua vao blacklist) va xoa refresh token.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Dang xuat",
                "responses": {
                    "200": {
                        "description": "Dang xuat thanh cong",
                        "schema": {
                            "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Dung refresh token hop le de cap cap access token + refresh token moi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Lam moi token",
                "parameters": [
                    {
                        "description": "Refresh request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "refresh_token": {
                                    "type": "string"
                                },
                                "user_id": {
                                    "type": "integer"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lam moi token thanh cong",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/go-familytree_internal_service.TokenPair"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Tao tai khoan nguoi dung moi bang email, mat khau va ten hien thi.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Dang ky tai khoan moi",
                "parameters": [
                    {
                        "description": "Registration info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/go-familytree_internal_service.RegisterInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tao tai khoan thanh cong",
                        "schema": {
                            "$ref": "#/definitions/go-familytree_pkg_response.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "go-familytree_internal_service.LoginInput": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "go-familytree_internal_service.RegisterInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                }
            }
        },
...
        "go-familytree_internal_service.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "description": "seconds",
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "go-familytree_pkg_response.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}